+++ src/test/java/com/jnape/palatable/lambda/monad/transformer/builtin/MaybeTTest.java
@@ -22,14 +22,18 @@ import static com.jnape.palatable.lambda.adt.Either.left;
 import static com.jnape.palatable.lambda.adt.Either.right;
 import static com.jnape.palatable.lambda.adt.Maybe.just;
 import static com.jnape.palatable.lambda.adt.Maybe.nothing;
+import static com.jnape.palatable.lambda.adt.Unit.UNIT;
 import static com.jnape.palatable.lambda.functions.builtin.fn2.GT.gt;
 import static com.jnape.palatable.lambda.functions.builtin.fn2.LT.lt;
 import static com.jnape.palatable.lambda.functor.builtin.Identity.pureIdentity;
 import static com.jnape.palatable.lambda.functor.builtin.Lazy.lazy;
 import static com.jnape.palatable.lambda.io.IO.io;
-import static com.jnape.palatable.lambda.monad.transformer.builtin.MaybeT.*;
+import static com.jnape.palatable.lambda.monad.transformer.builtin.MaybeT.liftMaybeT;
+import static com.jnape.palatable.lambda.monad.transformer.builtin.MaybeT.maybeT;
+import static com.jnape.palatable.lambda.monad.transformer.builtin.MaybeT.pureMaybeT;
 import static com.jnape.palatable.traitor.framework.Subjects.subjects;
 import static org.junit.Assert.assertEquals;
+import static testsupport.assertion.MonadErrorAssert.assertLaws;
 
 @RunWith(Traits.class)
 public class MaybeTTest {
@@ -39,6 +43,13 @@ public class MaybeTTest {
         return subjects(maybeT(right(just(1))),
                         maybeT(right(nothing())),
                         maybeT(left("foo")));
+    }
+
+    @Test
+    public void monadError() {
+        assertLaws(subjects(maybeT(new Identity<>(nothing())), maybeT(new Identity<>(just(1)))),
+                   UNIT,
+                   e -> maybeT(new Identity<>(just(2))));
     }
 
     @Test