+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - `IterateT#unfold` now only computes a single `Pure` for the given input
 - `ReaderT#fmap` and `StateT#fmap` avoid unnecessary calls to `pure`
+- `MaybeT` implements `MonadError`
 
 ### Added
 - `$`, function application represented as a higher-order `Fn2`